Tile-Based Movement, Player Animations, Object Collisions, and Enemy Encounters Base Logic
Implemented code logic for basic tile-based movement, for stopping player movement on certain objects, and for enemy encounters, as well as created the basic structure for the game world through tilemaps and tile palettes and blend trees to create idle and walking player movement animations.

Encountered minimal challenges as I followed Game Dev Experiement's "Make a Game Like Pokemon in Unity" YouTube series, though had to troublshoot my art sprites for a while as I transferred the project to a new computer and some of the sprites got corrupted. As a preface, I will be following the YouTube series to setup the basic mechanics of the game as I don't have significant experience in Unity, though I will be heavily modifying and revising it once I've implemented the necessary features.