Battle System Improvements, NPCs and Dialogue Systems, Overhauled Custom Animation Systems, and NPC Walk Behaviour
Improved the battle system by adding "Priority" parameters that determine whether a player or enemy move is initiated before the other. This has been designed so that it's easily modifiable in the Inspector for each move.

Furthermore, I implemented NPC code logic, allowing the player to interact with NPCs through dialogue when they press Z and face their direction, as well as allowing for modifiable pathing for the NPC. As there will be multiple NPCs present in the game, I use interfaces to store all of the common NPC functions so that I don't have to repeat code for each and every single NPC, improving performance.

Minimal challenges faced, though I had frequent logic errors when troublshooting the NPC and player interaction as I assigned the NPC and player with the same variable name of "Character". As a result, the Player object was displaying the same scripts and settings as the NPC.

Boss battles to be implemented in the next commit.