UI Scrolling, Pokemon Inventory Select, and Items
Implemented inventory scroll through arrow keys instead of through mouse movements, as well as implemented dynamic scrolling in the code. For example, if the list exceeds the maximum number of items viewable in an inventory list, arrows will be displayed on the top and/or bottom depending on where the player is hovering.

Implemented select and use functions for items in the inventory UI, allowing the player to switch to the party screen and select a Pokemon to use and apply the selected item.