[25/5/24] - Main Menu System and Partial Gameworld Setup
Implemented the main menu system, allowing the player to Play, view the Controls and return to the main menu, and quit the application. Code logic has also been applied allowing for SFX and timed events to be played correspondingly depending on the menu button pressed.

Also began the setup of the gameworld, wherein I rearranged the first and second scenes to align with my intended game design instead of their original template layouts.