Opening Crawl, Opening Cutscene Setup, Main Menu Revision, and Partial Audio Re-Implementation
Implemented the game's opening crawl, the setup for its opening cutscene, and revised the main menu with additional code logic that addressed some logic errors present such as the play button not being highlighted when opening the main menu.

Also replaced the game's audio with .mp3 instead of their original .wav files in order to increase the game's performance.